requirements.txt
requests==2.32.5
pandas==2.3.2
flask==3.1.2
Markdown==3.8.2
Pygments==2.19.2
python-dotenv==1.1.1
kiteconnect==5.0.1
